added docs about parry(especially RDATA)
mafia-delustion.txt concerns RDATA.
The state transition graph of mafia delusion is mafia.eps.
It was drew by igraph. The script is mafiadelusion.r
some bug report of parry is maybebugs.txt

cparry/t.lisp is an utility for show the wide view of
the property list structure in RDATA.

	new file:   ../cparry/t.lisp
	new file:   dedges.csv
	new file:   mafia-delusion.txt
	new file:   mafia.eps
	new file:   mafia.txt
	new file:   mafiadelusion.r
	new file:   maybebugs.txt
	new file:   pedges.csv